requirements.txt
streamlit
pandas
numpy
requests
yfinance
openpyxl
python-dotenv
firebase-admin
textblob
